requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
websockets>=12.0
websocket-client>=1.7.0
solders>=0.19.0
base58>=2.1.1
colorama>=0.4.6
pandas>=2.1.0
numpy>=1.26.0
requests>=2.31.0
scipy>=1.11.0
arch>=6.2.0
pandas-ta>=0.3.14b0
brotli>=1.1.0